fix(apps/client): HASHI-114 CI 테스트 실패 수정
