+++ apps/client/src/features/review/api/deleteReview.ts
@@ -1,4 +1,4 @@
-import { request } from '@/shared/api'
+import { request } from '@/shared/api/request'
 
 export const deleteReview = (reviewId: number) =>
   request<null>(`api/v1/reviews/${reviewId}`, { method: 'delete' })
