+++ apps/client/src/features/review/queries/useMyReviewCountQuery.test.tsx
@@ -3,13 +3,13 @@ import { renderHook, waitFor } from '@testing-library/react'
 import type { PropsWithChildren } from 'react'
 import { describe, expect, it, vi } from 'vitest'
 
-import { getMyReviewCount } from '@/features/review/api'
+import { getMyReviewCount } from '@/features/review/api/getMyReviewCount'
 import {
   myReviewQueryKeys,
   useMyReviewCountQuery,
 } from '@/features/review/queries'
 
-vi.mock('@/features/review/api', () => ({
+vi.mock('@/features/review/api/getMyReviewCount', () => ({
   getMyReviewCount: vi.fn(),
 }))
