+++ apps/client/src/pages/reviewDetail/api/getMyReviewDetail.ts
@@ -1,5 +1,5 @@
 import type { components } from '@/shared/api/generated/openapi'
-import { request } from '@/shared/api'
+import { request } from '@/shared/api/request'
 
 export type MyReviewDetailData = components['schemas']['MyReviewDetailResponse']
