+++ apps/client/src/pages/reviewDetail/queries/useReviewDetailQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query'
 
-import { myReviewQueryKeys } from '@/features/review/queries'
+import { myReviewQueryKeys } from '@/features/review/queries/myReviewQueryKeys'
 import { getMyReviewDetail } from '@/pages/reviewDetail/api/getMyReviewDetail'
 
 export const useReviewDetailQuery = (reviewId: number | null) =>
